+++ Magic/CHANGELOG.md
@@ -44,6 +44,7 @@
  - Added Holy Light spell to Wizard path
  - Add "equation" parameter to Command action, works just like Message action
  - Add "enable_vanilla_sounds" configuration option for those of you who *really* hate the new sound configs
+ - Add `critical_probability` and `crticail_damage_multiplier` parameters to Damage action
 
 # 8.2.3